// src/interfaces/Glassix/Ticket.ts
export interface Participant {
    id: number
    name: string
    type: string
    identifier: string
    departmentIdentifier: string
    displayName: string
    protocolType: ProtocolType
    subProtocolType: SubProtocolType
    isActive: boolean
    isDeleted: boolean
    contactId: string
    userName: string
}

export interface Transaction {
    guidTransactionId: string
    id: number
    dateTime: Date
    fromProtocolType: ProtocolType
    type: string
    text: string
    html: string
    status: string
    fromParticipant: Participant
    attachments: any[]
    payload?: Payload
}


export interface Payload {
    templates: Template[]
    enableFreeTextInput: boolean
    quickReplies?: any[]
}

export interface Template {
    title: string
    titleHtml: string
    subtitle: string
    subtitleHtml: string
    coverImageHeight: number
    coverImageWidth: number
    buttons: Button[]
}
// ...
export enum ProtocolType {
    SMS = 'SMS',
    WhatsApp = 'WhatsApp',
    Mail = 'Mail',
    Web = 'Web',
    FacebookMessenger = 'FBmessenger',
    WebChatViaSMSLink = 'WebViaSMS',
    FacebookFeed = 'FacebookFeed',
    InstagramFeed = 'InstagramFeed',
    InstagramDirect = 'InstagramDM',
    PhoneCall = 'PhoneCall',
    Viber = 'Viber',
    Twitter = 'Twitter',
    AppleBusinessChat = 'AppleBusinessChat',
    GoogleBusinessMessages = 'GoogleBusinessMessages',
    GoogleBusinessReviews = 'Google Business Reviews'
}

export enum SubProtocolType {
    MailTo = 'MailTo',
    MailCc = 'MailCc',
    MailBcc = 'MailBcc'
}